llm_services/requirements.txt
fastapi
httpx
python-dotenv
boto3
scikit-learn
uvicorn
mistralai
python-multipart
